+++ drivers/gpu/nvgpu/gm20b/regops_gm20b.h
@@ -2,7 +2,7 @@
  *
  * Tegra GK20A GPU Debugger Driver Register Ops
  *
- * Copyright (c) 2013-2017, NVIDIA CORPORATION. All rights reserved.
+ * Copyright (c) 2013-2018 NVIDIA CORPORATION. All rights reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a
  * copy of this software and associated documentation files (the "Software"),
@@ -28,17 +28,17 @@
 struct dbg_session_gk20a;
 
 const struct regop_offset_range *gm20b_get_global_whitelist_ranges(void);
-int gm20b_get_global_whitelist_ranges_count(void);
+u64 gm20b_get_global_whitelist_ranges_count(void);
 const struct regop_offset_range *gm20b_get_context_whitelist_ranges(void);
-int gm20b_get_context_whitelist_ranges_count(void);
+u64 gm20b_get_context_whitelist_ranges_count(void);
 const u32 *gm20b_get_runcontrol_whitelist(void);
-int gm20b_get_runcontrol_whitelist_count(void);
+u64 gm20b_get_runcontrol_whitelist_count(void);
 const struct regop_offset_range *gm20b_get_runcontrol_whitelist_ranges(void);
-int gm20b_get_runcontrol_whitelist_ranges_count(void);
+u64 gm20b_get_runcontrol_whitelist_ranges_count(void);
 const u32 *gm20b_get_qctl_whitelist(void);
-int gm20b_get_qctl_whitelist_count(void);
+u64 gm20b_get_qctl_whitelist_count(void);
 const struct regop_offset_range *gm20b_get_qctl_whitelist_ranges(void);
-int gm20b_get_qctl_whitelist_ranges_count(void);
+u64 gm20b_get_qctl_whitelist_ranges_count(void);
 int gm20b_apply_smpc_war(struct dbg_session_gk20a *dbg_s);
 
 #endif /* __REGOPS_GM20B_H_ */
