+++ drivers/gpu/nvgpu/gp106/regops_gp106.h
@@ -2,7 +2,7 @@
  *
  * Tegra GP106 GPU Debugger Driver Register Ops
  *
- * Copyright (c) 2016-2017, NVIDIA CORPORATION. All rights reserved.
+ * Copyright (c) 2016-2018, NVIDIA CORPORATION. All rights reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a
  * copy of this software and associated documentation files (the "Software"),
@@ -26,17 +26,17 @@
 #define __REGOPS_GP106_H_
 
 const struct regop_offset_range *gp106_get_global_whitelist_ranges(void);
-int gp106_get_global_whitelist_ranges_count(void);
+u64 gp106_get_global_whitelist_ranges_count(void);
 const struct regop_offset_range *gp106_get_context_whitelist_ranges(void);
-int gp106_get_context_whitelist_ranges_count(void);
+u64 gp106_get_context_whitelist_ranges_count(void);
 const u32 *gp106_get_runcontrol_whitelist(void);
-int gp106_get_runcontrol_whitelist_count(void);
+u64 gp106_get_runcontrol_whitelist_count(void);
 const struct regop_offset_range *gp106_get_runcontrol_whitelist_ranges(void);
-int gp106_get_runcontrol_whitelist_ranges_count(void);
+u64 gp106_get_runcontrol_whitelist_ranges_count(void);
 const u32 *gp106_get_qctl_whitelist(void);
-int gp106_get_qctl_whitelist_count(void);
+u64 gp106_get_qctl_whitelist_count(void);
 const struct regop_offset_range *gp106_get_qctl_whitelist_ranges(void);
-int gp106_get_qctl_whitelist_ranges_count(void);
+u64 gp106_get_qctl_whitelist_ranges_count(void);
 int gp106_apply_smpc_war(struct dbg_session_gk20a *dbg_s);
 
 #endif /* __REGOPS_GP106_H_ */
